Refactor User Pages into separate files.
