+++ src/App.js
@@ -8,18 +8,17 @@ import loginPage from './pages/LoginPage'
 
 import { CourseList, CourseEdit, CourseCreate } from './pages/courses';
 import { ReferenceExamplesList, ReferenceExamplesEdit, ReferenceExamplesCreate } from './pages/referenceExamples';
-import { UserList, UserEdit, UserCreate } from './pages/users';
 import { SnapshotList, SnapshotShow } from './pages/snapshots';
 import { NotifList, NotifEdit, NotifCreate } from './pages/Notifications';
 import { CollectionList, CollectionShow } from './pages/Collections';
 
 import scenes from './pages/scene';
 import googleLogins from "./pages/googleLogin";
+import users from "./pages/user";
 
 import {
   List                 as CourseIcon,
   Help                 as RefExIcon,
-  SupervisedUserCircle as UserIcon,
   CameraEnhance        as SnapshotIcon,
   Filter               as CollectionIcon,
   PriorityHigh         as NotificationIcon,
@@ -27,14 +26,14 @@ import {
 
 const App = () => (
   <Admin dashboard={Dashboard} loginPage={loginPage} authProvider={authProvider} dataProvider={dataProvider}>
-    <Resource name="scenes" {...scenes} />
-    <Resource name="collections" list={CollectionList} show={CollectionShow} icon={CollectionIcon} />
-    <Resource name="courses" list={CourseList} edit={CourseEdit} create={CourseCreate} icon={CourseIcon} />
-    <Resource name="referenceExamples" list={ReferenceExamplesList} edit={ReferenceExamplesEdit} create={ReferenceExamplesCreate} icon={RefExIcon} />
-    <Resource name="users" list={UserList} edit={UserEdit} create={UserCreate} icon={UserIcon} />
-    <Resource name="googlelogins"  {...googleLogins} />
-    <Resource name="notifications" list={NotifList} edit={NotifEdit} create={NotifCreate} icon={NotificationIcon} />
-    <Resource name="snapshots" list={SnapshotList} show={SnapshotShow} icon={SnapshotIcon} />
+    <Resource name="scenes"           {...scenes} />
+    <Resource name="collections"      list={CollectionList} show={CollectionShow} icon={CollectionIcon} />
+    <Resource name="courses"          list={CourseList} edit={CourseEdit} create={CourseCreate} icon={CourseIcon} />
+    <Resource name="referenceExamples"list={ReferenceExamplesList} edit={ReferenceExamplesEdit} create={ReferenceExamplesCreate} icon={RefExIcon} />
+    <Resource name="users"            {...users} />
+    <Resource name="googlelogins"     {...googleLogins} />
+    <Resource name="notifications"    list={NotifList} edit={NotifEdit} create={NotifCreate} icon={NotificationIcon} />
+    <Resource name="snapshots"        list={SnapshotList} show={SnapshotShow} icon={SnapshotIcon} />
   </Admin>
 );
