+++ src/pages/user/index.js
@@ -1,0 +1,12 @@
+import UserIcon from "@mui/icons-material/SupervisedUserCircle";
+
+import UserList from "./UserList";
+import UserEdit from "./UserEdit";
+import UserCreate from "./UserCreate"; 
+
+export default {
+    icon: UserIcon,
+    list: UserList,
+    edit: UserEdit,
+    create: UserCreate
+};
